+++ src/Equinox.Cosmos/Cosmos.fs
@@ -1078,10 +1078,23 @@ module Caching =
             (slidingExpiration : TimeSpan)
             (category : ICategory<'event, 'state, Container*string, 'context>)
             : ICategory<'event, 'state, Container*string, 'context> =
-        let mkCacheEntry (initialToken : StreamToken, initialState : 'state) = new CacheEntry<'state>(initialToken, initialState, Token.supersedes)
+        let mkCacheEntry (initialToken : StreamToken, initialState : 'state) = CacheEntry<'state>(initialToken, initialState, Token.supersedes)
         let options = CacheItemOptions.RelativeExpiration slidingExpiration
         let addOrUpdateSlidingExpirationCacheEntry streamName value = cache.UpdateIfNewer(prefix + streamName, options, mkCacheEntry value)
         CategoryTee<'event, 'state, 'context>(category, addOrUpdateSlidingExpirationCacheEntry) :> _
+
+    let applyCacheUpdatesWithFixedTimeSpan
+            (cache : ICache)
+            (prefix : string)
+            (lifetime : TimeSpan)
+            (category : ICategory<'event, 'state, Container*string, 'context>)
+            : ICategory<'event, 'state, Container*string, 'context> =
+        let mkCacheEntry (initialToken : StreamToken, initialState : 'state) = CacheEntry<'state>(initialToken, initialState, Token.supersedes)
+        let addOrUpdateFixedLifetimeCacheEntry streamName value =
+            let expirationPoint = let creationDate = DateTimeOffset.UtcNow in creationDate.Add lifetime
+            let options = CacheItemOptions.AbsoluteExpiration expirationPoint
+            cache.UpdateIfNewer(prefix + streamName, options, mkCacheEntry value)
+        CategoryTee<'event, 'state, 'context>(category, addOrUpdateFixedLifetimeCacheEntry) :> _
 
 type private Folder<'event, 'state, 'context>
     (   category: Category<'event, 'state, 'context>, fold: 'state -> 'event seq -> 'state, initial: 'state,
@@ -1155,6 +1168,10 @@ type CachingStrategy =
     /// NB while a strategy like EventStore.Caching.SlidingWindowPrefixed is obviously easy to implement, the recommended approach is to
     /// track all relevant data in the state, and/or have the `unfold` function ensure _all_ relevant events get held in the `u`nfolds in tip
     | SlidingWindow of ICache * window: TimeSpan
+    /// Retain a single 'state per streamName, together with the associated etag
+    /// Upon expiration of the span, a reload is triggered
+    /// Typically combined with `Equinox.ResolveOption.AllowStale` to minimize loads
+    | FixedTimeSpan of ICache * span: TimeSpan
 
 [<NoComparison; NoEquality; RequireQualifiedAccess>]
 type AccessStrategy<'event,'state> =
@@ -1197,7 +1214,8 @@ type Resolver<'event, 'state, 'context>
     let readCacheOption =
         match caching with
         | CachingStrategy.NoCaching -> None
-        | CachingStrategy.SlidingWindow(cache, _) -> Some(cache, null)
+        | CachingStrategy.SlidingWindow (cache, _)
+        | CachingStrategy.FixedTimeSpan (cache, _) -> Some(cache, null)
     let isOrigin, mapUnfolds =
         match access with
         | AccessStrategy.Unoptimized ->                      (fun _ -> false), Choice1Of3 ()
@@ -1211,8 +1229,10 @@ type Resolver<'event, 'state, 'context>
     let category : ICategory<_, _, Container*string, 'context> =
         match caching with
         | CachingStrategy.NoCaching -> folder :> _
-        | CachingStrategy.SlidingWindow(cache, window) ->
+        | CachingStrategy.SlidingWindow (cache, window) ->
             Caching.applyCacheUpdatesWithSlidingExpiration cache null window folder
+        | CachingStrategy.FixedTimeSpan (cache, span) ->
+            Caching.applyCacheUpdatesWithFixedTimeSpan cache null span folder
 
     let resolveStream (streamId, maybeContainerInitializationGate) opt context =
         { new IStream<'event, 'state> with